Create class Date
Date works with 12 different formats.
The output of the date can be formatted to a specific format
using one of 12 manipulators.
Date can be inputted in one of the formats chosen by the user.

Operator- counts difference in days between two dates.
Operator int() converts date to amount of days passed after 00-00-0000.

The full range of comparison operators is implemented (==, !=, >, <, >=, <=).

When creating a date, month can be specified either as a number
or as a enum element (to make it more readable).

Signed-off-by: Polina Shlepakova <polinashlepakova@gmail.com>